user login, logout, login is required

+++ firstproject/baseapp/views.py
@@ -3,6 +3,11 @@ from django.db.models import Q
 from django.http import HttpResponse
 from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+
 
 
 # Create your views here.
@@ -12,6 +17,37 @@ rooms = [
        {"id":2, "name": "learn python"},
        {"id":3, "name": "learn full stack"}
 ]
+
+def loginPage(request):
+    if request.user.is_authenticated:
+       return redirect('home')
+
+    if request.method == 'POST':
+       print("post request:", request.POST)
+       username = request.POST.get('usernamee')
+       password = request.POST.get('passwordd')
+       print("username:", username, "password:", password)
+
+       try:
+           user = User.objects.get(username = username)
+       except:
+            messages.error(request, "User Doesn't exist")
+
+       user = authenticate(request, username=username, password=password)
+
+       if user is not None:
+          login(request, user)
+          return redirect('home')
+       else:
+           messages.error(request, "Username or Password are does not exists")
+           
+           
+    context = {}
+    return render(request, 'baseapp/login_register.html', context)
+    
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -26,6 +62,7 @@ def room(request, pk):
     context = {"rumm": room}
     return render(request, 'baseapp/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -38,9 +75,14 @@ def createRoom(request):
     context = {"form": form}
     return render(request, 'baseapp/create_room.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id = int(pk))
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+       return HttpResponse("You are not allowed")
+
     if request.method == 'POST':
        form = RoomForm(request.POST, instance=room)
        if form.is_valid():
@@ -51,8 +93,12 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'baseapp/create_room.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id = int(pk))
+    if request.user != room.host:
+       return HttpResponse("You are not allowed")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
